Fix penalty
Previous implementation of penalty is flawed.

Previously negative penalty is added.
However, it can be the case where one worker exceed the time
by only a little bit and all the other workers are under
time by a lot.

The penalty will still be negative in this case,
even though a worker has exceeded the time.

Fix the penalty to only include those that exceed the time.